refactor: Extract downloadPath creation to function
