+++ src/lib/component/SaveAnnotationDialog/bind/downloadAnnotationFile/createDownloadPathByFormat.js
@@ -1,0 +1,13 @@
+import createDownloadPath from '../../../createDownloadPath'
+import convertJSONAnnotationToInline from '../../../../convertJSONAnnotationToInline'
+
+export default async function createDownloadPathByFormat(data, format) {
+  if (format === 'json') {
+    return createDownloadPath(data)
+  } else if (format === 'inline') {
+    const inlineData = await convertJSONAnnotationToInline(data)
+
+    const blob = new Blob([inlineData], { type: 'text/plain' })
+    return URL.createObjectURL(blob)
+  }
+}
